% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{activity_corr_table}
\alias{activity_corr_table}
\title{Correspondence table describing the business sectors hierarchy.}
\format{
A data frame with 92 rows and 3 variables:
\describe{
  \item{A10}{business sectors in 10 categories}
  \item{A21}{business sectors in 21 categories}
  \item{A88}{business sectors in 88 categories}
}
}
\usage{
activity_corr_table
}
\description{
A dataset describing the nesting of three levels of business sectors, useful
when working with the ACTIVITY variables in the turnover_ datasets.
}
\details{
Use the \code{write_hrc2} function to create a .hrc file from this
correspondence table.
}
\keyword{datasets}
