% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_tab_rtauargus.R
\name{auto_limit}
\alias{auto_limit}
\title{Automatically computes an intermediate size limit for table reduction}
\usage{
auto_limit(dfs, totcode, hrcfiles = NULL)
}
\arguments{
\item{dfs}{data.frame containing 4 or 5 categorical variables}

\item{totcode}{named vector of totals for categorical variables}

\item{hrcfiles}{named vector of hrc files (optional)}
}
\value{
An integer representing the computed row limit.
}
\description{
Uses \code{explore_reduce_dims()} to find the median configuration in terms of
generated tables, and returns the minimum \code{max_size} among the configurations
having that number of tables.
}
