% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfs5_agg}
\alias{dfs5_agg}
\title{Aggregated 5-dimensional example dataset}
\format{
A \code{data.frame} (or \code{data.table}) containing combinations of the 5 categorical variables and cell counts.
}
\source{
Example dataset generated for package documentation and testing.
}
\usage{
dfs5_agg
}
\description{
An example dataset containing a 5-dimensional cross-tabulated table,
notably used in the \code{split_tab} vignette.
}
\details{
The table includes the following categorical variables:
\describe{
\item{ACT}{Activity (25 modalities: 1 Total + 4 groups + 20 sub-levels)}
\item{GEO}{Geography (36 modalities: 1 Total + 5 regions + 30 sub-regions)}
\item{SEX}{Sex / Category (11 modalities: 1 Total + 10 detailed categories)}
\item{AGE}{Age group (16 modalities: 1 Total + 15 detailed categories)}
\item{ECO}{Economic sector (3 modalities: 1 Total + 2 detailed categories)}
\item{N}{Number of observations / cell frequency}
}
}
\seealso{
\code{\link{tot5}}, \code{\link{hrc5}}
}
\keyword{datasets}
