% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enterprise_template}
\alias{enterprise_template}
\title{Extract of SBS Eurostat template.}
\format{
A data frame with 3 168 rows and 5 variables:
\describe{
\item{TIME_PERIOD}{Integer. The year of observation (e.g., 2022).}
\item{INDICATOR}{Character. The type of indicator (e.g., "SAL").}
\item{ACTIVITY}{Character. The sector of activity coded using a standard classification (e.g., "B").}
\item{NUMBER_EMPL}{Character. Employment size category (e.g., "E0", "E1T4", "E5T9", "EGE10").}
\item{LEGAL_FORM}{Character. Legal form of the enterprise, where "_T" represents all legal forms combined.}
}
}
\usage{
enterprise_template
}
\description{
A dataset describing all the cells expected by Eurostat.
}
\keyword{datasets}
