% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_var_to_merge.R
\name{explore_reduce_dims}
\alias{explore_reduce_dims}
\title{Analytically computes all possible table splits and summarizes their statistics.}
\usage{
explore_reduce_dims(dfs, totcode, hrcfiles = NULL)
}
\arguments{
\item{dfs}{data.frame containing 4 or 5 categorical variables}

\item{totcode}{named vector of totals for categorical variables}

\item{hrcfiles}{named vector of hrc files (optional)}
}
\value{
A deduplicated data.frame containing the following columns:
\itemize{
\item \code{nb_tab}: number of generated tables
\item \code{nb_hrc}: number of remaining hierarchical variables
\item \code{min_size}: minimum table size (rows)
\item \code{med_size}: median table size (rows, rounded to integer)
\item \code{max_size}: maximum table size (rows)
}
}
\description{
Analytically computes all possible table splits and summarizes their statistics.
}
