% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_template.R
\name{format_template}
\alias{format_template}
\title{Determines the tables described in a template gathering all the published cells}
\usage{
format_template(data, indicator_column, spanning_var_tot, field_columns)
}
\arguments{
\item{data}{template gathering all the published cells}

\item{indicator_column}{name of the column in which the indicators are}

\item{spanning_var_tot}{a named list of the spanning variables and their totals}

\item{field_columns}{vector of all the columns that are fields (ex: year of collect)}
}
\value{
named list of a dataframe describing the tables (metadata) and a list of
the modalities of each hierarchical variable (modalities)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
data(enterprise_template)

template_formatted <- format_template(
  data = enterprise_template,
  indicator_column = "INDICATOR",
  spanning_var_tot = list(
    ACTIVITY = "BTSXO_S94",
    NUMBER_EMPL = "_T",
    LEGAL_FORM = "_T"),
  field_columns = c("TIME_PERIOD")
)

}
