% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_template.R
\name{get_combinations}
\alias{get_combinations}
\title{Generate All Combinations of spanning variables}
\usage{
get_combinations(criteria)
}
\arguments{
\item{criteria}{A named list or vector, where the names represent the keys.}
}
\value{
A list of combinations, where each combination is a character vector of key names.
}
\description{
This function generates all possible combinations of keys from a given named list or vector.
}
\examples{
\dontrun{
criteria <- list(key1 = "value1", key2 = "value2", key3 = "value3")
get_combinations(criteria)
# Returns a list of combinations: c("key1"), c("key2"), c("key3"), c("key1", "key2"), ...
}
}
