% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_tab_names.R
\name{grp_tab_names}
\alias{grp_tab_names}
\title{Group Tables Based on Inclusion Relationships}
\usage{
grp_tab_names(list_split)
}
\arguments{
\item{list_split}{A list of data frames, where each data frame describes the
inclusion relationships (\code{from} and \code{to}) between tables in a cluster.
Typically, this is the output of the \code{create_edges()} function.}
}
\value{
A list of results for each cluster, where each result is a list
containing:
\itemize{
\item \code{tab_finales}: A data frame describing the final relationships between
grouped tables. It includes only the tables necessary for protection.
\item \code{passage_nom_tab}: A data frame mapping original table names (\code{Original})
to their respective groups (\code{Group}).
}
}
\description{
This function regroups tables that are included in each other into clusters,
keeping only the tables necessary for protection. Based on the inclusion
relationships detected by the \code{create_edges()} function, the tables are
aggregated to minimize redundancy. The output identifies the final tables
that need to be protected and provides a mapping of original table names
to their respective groups.
}
\examples{
# Example data
data(metadata_pizza_lettuce)

# Convert wide metadata to long format
metadata_pizza_lettuce_long <- wide_to_long(metadata_pizza_lettuce)

# Identify hierarchical relationships
list_hrc_identified <- identify_hrc(metadata_pizza_lettuce_long)

# Split tables into clusters
list_split <- split_in_clusters(list_hrc_identified)

# Detect inclusion relationships
list_desc_links <- create_edges(list_split)

# Group tables based on inclusion relationships
list_translation_tables <- grp_tab_names(list_desc_links)

# View structure of the results
str(list_translation_tables)

}
