% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hrc5}
\alias{hrc5}
\title{Hierarchies for the \code{dfs5_agg} dataset variables}
\format{
A named vector or list specifying the hierarchy for each variable.
}
\usage{
hrc5
}
\description{
Specification of hierarchy files (\code{.hrc}) or structures associated
with the hierarchical variables of the \code{\link{dfs5_agg}} table (notably \code{ACT} and \code{GEO}).
}
\seealso{
\code{\link{dfs5_agg}}, \code{\link{tot5}}
}
\keyword{datasets}
