% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_hrc_with_eq.R
\name{identify_hrc_with_eq}
\alias{identify_hrc_with_eq}
\title{Rename variables based on their hierarchies and their equations}
\usage{
identify_hrc_with_eq(df_metadata_long, df_eq_indicator)
}
\arguments{
\item{df_metadata_long}{A data frame in long format with the following
required columns:
\itemize{
\item \code{table_name}: Identifies the table.
\item \code{field} : name of the field of the table.
\item \code{indicator}: name of the indicator of the table.
\item \code{hrc_indicator}: Specifies linked response variables.
\item \verb{spanning_*}, \verb{hrc_spanning_*}: Spanning variables and their hierarchies.
}}

\item{df_eq_indicator}{A dataframe containing the equations on indicators with
the following required columns :
\itemize{
\item \code{eq_name}: Name of the equation.
\item \code{eq_indicator}: The equation for example, A = B + C.
\item \code{unit}: The unit of the indicators in the equation.
}}
}
\value{
\code{list_hrc_identified}, a list with two elements:
\itemize{
\item \code{df_indicators}: The updated data frame with renamed variables and grouped
response variables.
\item \code{df_variable_info}: A data frame mapping original variable names
(\code{spanning_old}) to their renamed counterparts (\code{spanning}).
}
}
\description{
This function renames variables in a long-format metadata data frame based on
their hierarchical groupings. Spanning variables are renamed using the name
of their hierarchy in uppercase, while response variables linked by an
equation (specified in the \code{hrc_indicator} column) are grouped together, and
a new grouping variable is created.
}
\examples{
\dontrun{
data(metadata_pizza_lettuce)

metadata_pizza_lettuce_long <- wide_to_long(metadata_pizza_lettuce)
 df_eq_ex <- data.frame(
 eq_name = c("eq1"),
 eq_indicator = c("ca_salades = ca_batavia + ca_mache"),
 unit = c("EUR"))

list_hrc_identified <- identify_hrc(metadata_pizza_lettuce_long, df_eq_ex)

str(list_hrc_identified)
}

}
