% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_var_to_merge.R
\name{import_hierarchy}
\alias{import_hierarchy}
\title{Import a hierarchy and cache its levels}
\usage{
import_hierarchy(hrcfile, total)
}
\arguments{
\item{hrcfile}{character path to a .hrc file}

\item{total}{character, the root total for the hierarchy}
}
\value{
a list of character vectors, each representing a node of the hierarchy
}
\description{
Import a hierarchy and cache its levels
}
\keyword{internal}
