% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indiv_dt}
\alias{indiv_dt}
\title{Companies data at individual level.}
\format{
A data.table with 9 786 rows and 12 variables:
\describe{
\item{A10}{business sector, not hierarchical}
\item{A21}{business sector, not hierarchical but nested in A10}
\item{A88}{business sector, not hierarchical but nested in A21}
\item{CJ}{legal category, not hierarchical}
\item{TYPE}{type of distribution, not hierarchical}
\item{SIZE}{Number of employees (categorical), not hierarchical}
\item{NUTS1}{NUTS 1 level of European administrative regions, not hierarchical}
\item{NUTS2}{NUTS 2 level of European administrative regions, not hierarchical}
\item{NUTS3}{NUTS 3 level of European administrative regions, not hierarchical}
\item{WEIGHT}{Weight of the companies, numeric}
\item{TURNOVER}{Turnover, numeric}
\item{PRODUCTION}{Production, numeric}
}
}
\usage{
indiv_dt
}
\description{
Companies data at individual level.
}
\keyword{datasets}
