% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_pizza_lettuce}
\alias{metadata_pizza_lettuce}
\title{Metadata for pizza and lettuce dataset.}
\format{
A data frame with 12 rows and 9 variables:
\describe{
\item{table_name}{Character. Name of the table (e.g., "T1", "T2").}
\item{field}{Character. Data source or category (e.g., "france_entreprises_2023").}
\item{hrc_field}{Logical. Indicates if hierarchical classification is applied to the field (NA if not applicable).}
\item{indicator}{Character. The indicator being measured (e.g., "to_pizza").}
\item{hrc_indicator}{Character. Hierarchical classification for the indicator (NA if not applicable).}
\item{spanning_1}{Character. First spanning variable (e.g., "nuts2", "nuts3", "a10").}
\item{hrc_spanning_1}{Character. Hierarchical classification for spanning_1 (e.g., "hrc_nuts", "hrc_naf").}
\item{spanning_2}{Character. Second spanning variable (e.g., "size", "nuts2", "nuts3").}
\item{hrc_spanning_2}{Character. Hierarchical classification for spanning_2 (NA if not applicable).}
}
}
\usage{
metadata_pizza_lettuce
}
\description{
A dataset containing metadata information for various tables and indicators related to
enterprises turnover selling pizzas and lettuces in France.
}
\keyword{datasets}
