% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_var_to_merge.R
\name{nb_nodes}
\alias{nb_nodes}
\title{Count the number of nodes in a hierarchical file}
\usage{
nb_nodes(
  hrcfiles = NULL,
  v = NULL,
  hrc_name = TRUE,
  totcode = NULL,
  total = NULL
)
}
\arguments{
\item{hrcfiles}{named vector of hrc files, can be NULL}

\item{v}{variable name (must be present in names(hrcfiles) if hrcfiles is not NULL)}

\item{hrc_name}{logical, if TRUE v is a name in hrcfiles, else hrcfiles is the hrc path itself (optional)}
}
\value{
integer number of nodes
}
\description{
Count the number of nodes in a hierarchical file
}
\keyword{internal}
