% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts23_fr_corr_table}
\alias{nuts23_fr_corr_table}
\title{data - Correspondence table describing the NUTS hierarchy.}
\format{
A data frame with 92 rows and 3 variables:
\describe{
  \item{NUTS2}{NUTS2 levels in France - equivalent of French "Régions"}
  \item{NUTS3}{NUTS3 levels in France - equivalent of French "Départements"}
}
}
\usage{
nuts23_fr_corr_table
}
\description{
A dataset describing the nesting of NUTS2 and NUTS3 levels for Metropolitan France, useful
when working with the NUTS variables in the turnover_ datasets.
}
\details{
Use the \code{write_hrc2} function to create a .hrc file from this
correspondence table.
}
\keyword{datasets}
