% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtauargus.R
\name{rtauargus}
\alias{rtauargus}
\title{Secrétise des tableaux à partir de microdonnées}
\usage{
rtauargus(microdata, explanatory_vars, safety_rules, suppress, ...)
}
\arguments{
\item{microdata}{[\strong{obligatoire}] data.frame contenant les
microdonnées.}

\item{explanatory_vars}{[\strong{obligatoire}] variables catégorielles, sous
forme de liste de vecteurs. Chaque élément de la liste est un vecteur des
noms des variables formant une tabulation.
Exemple : \code{list(c("CJ", "A21"), c("SEXE", "REGION"))} pour le premier
tableau croisant \code{CJ} x \code{A21} et le deuxième tableau croisant
\code{SEXE} x \code{REGION}.
Si une seule tabulation, un simple vecteur des variables à croiser est
accepté (pas besoin de \code{list(...)}).}

\item{safety_rules}{[\strong{obligatoire}] règle(s) de secret primaire (hors
pondération). Chaîne de caractères en syntaxe batch Tau-Argus.}

\item{suppress}{[\strong{obligatoire}] méthode(s) de gestion du secret
secondaire (syntaxe batch de Tau-Argus). Si la méthode est la même pour
chaque tabulation, le premier paramètre (numéro du tableau) sera ignoré et
renuméroté automatiquement.}

\item{...}{paramètres optionnels pour \code{micro_asc_rda}, \code{micro_arb}
et \code{run_tauargus}. Voir l'aide de ces fonctions.}
}
\value{
Si \code{import = TRUE}, une liste de data.frames (tableaux
  secrétisés), \code{NULL} sinon.
}
\description{
Secrétise des tableaux construits à partir de microdonnées et des
spécifications des croisements. La fonction permet d'effectuer le processus
complet, à savoir la création des fichiers asc et rda, la construction
du fichier arb, le lancement effectif de Tau-Argus et la récupération
éventuelle des résultats dans R.
}
\details{
La fonction exécute séquentiellement les fonctions : \itemize{
  \item{
    \code{\link{micro_asc_rda}} \code{->}
    \code{\link{micro_arb}} \code{->}
    \code{\link{run_tauargus}}
  }
}

Les fichiers intermédiaires sans nom renseigné (\code{asc_filename}...) seront
créés dans un dossier temporaire, avec des noms générés aléatoirement. Ce
mécanisme permet à l'utilisateur de s'abstraire de la préparation des données
propre à Tau-Argus et de maintenir l'intégralité de la chaîne de traitements
dans R.
}
\examples{
rtauargus(
  microdata = data.frame(V1 = c("A", "A", "B", "C", "A"), V2 = "Z"),
  explanatory_vars = c("V1", "V2"),
  safety_rules = "FREQ(3,10)",
  suppress = "GH(.,100)",
  output_options = "AS+" # (exemple de parametre optionnel pour micro_arb)
)
}
