% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauargus_exe.R
\name{set_tauargus_exe}
\alias{set_tauargus_exe}
\title{Spécifie l'emplacement de Tau Argus}
\usage{
set_tauargus_exe(path_exe = "Y:/Logiciels/Tau/TauArgus.exe", quiet = FALSE)
}
\arguments{
\item{path_exe}{chemin complet et nom du programme Tau Argus.}

\item{quiet}{\code{TRUE} pour opérer de manière silencieuse.}
}
\description{
Modifie l'option \code{rtauargus_exe}. Si le paramètre n'est pas renseigné,
revient à l'emplacement par défaut (version portable sur AUS).
}
\details{
La fonction ne vérifie pas si le chemin existe réellement ou si l'exécutable
est bien présent.
}
