% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_in_clusters.R
\name{split_in_clusters}
\alias{split_in_clusters}
\title{Split a Data Frame into Clusters of Linked Tables}
\usage{
split_in_clusters(list_hrc_identified)
}
\arguments{
\item{list_hrc_identified}{A list returned by the \code{identify_hrc} function. The first
element of the list must be a data frame containing the variables:
\itemize{
\item \code{field}: A grouping variable.
\item \code{hrc_field}: The hierarchical counterpart of \code{field}.
\item \code{indicator}: A variable used to link tables.
\item \code{hrc_indicator}: The hierarchical counterpart of \code{indicator}.
}}
}
\value{
A named list of data frames. Each element of the list corresponds to a
cluster of linked tables, and the names of the list elements reflect the
hierarchical grouping path.
}
\description{
This function splits a data frame into a list of data frames (or clusters), where
each cluster represents a group of linked tables. Tables are grouped based on the
\code{field} variable and their corresponding \code{indicator} values. The hierarchical
relationships specified in the \verb{hrc_*} columns are also considered.
}
\details{
The function handles cases where the \code{field} variable is constant across the data
frame or when \code{field} varies, in which case the data frame is split by \code{field} and
further split by \code{indicator} and \code{hrc_indicator}.
}
\examples{
library(dplyr)
data(metadata_pizza_lettuce)

metadata_pizza_lettuce_long <- wide_to_long(metadata_pizza_lettuce)
# Identify hierarchical relationships
list_hrc_identified <- identify_hrc(metadata_pizza_lettuce_long)
# Split into clusters
list_split <- split_in_clusters(list_hrc_identified)

# View the structure of the result
str(list_split)

}
