% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_arb.R
\name{tab_arb}
\alias{tab_arb}
\title{Creates a batch file (.arb) for tabular data in order to run Tau Argus}
\usage{
tab_arb(
  arb_filename = NULL,
  tab_filename,
  rda_filename = NULL,
  hst_filename = NULL,
  explanatory_vars,
  value = getOption("rtauargus.response_var"),
  safety_rules,
  suppress,
  separator = getOption("rtauargus.separator"),
  output_names = NULL,
  output_type = getOption("rtauargus.output_type"),
  output_options = getOption("rtauargus.output_options"),
  gointeractive = FALSE
)
}
\arguments{
\item{arb_filename}{path of the arb filename, if unspecified, creates a temporary
file \cr
(nom du fichier arb généré (avec extension). Si non renseigné, un fichier temporaire.)}

\item{tab_filename}{path of the tab_filename.\cr
(nom du fichier .tab (avec extension).)}

\item{rda_filename}{rda file name (with .rda extension) \cr
nom du fichier rda (avec extension)}

\item{hst_filename}{Apriori file name, syntax detailed below,
Example : hst_filename = "path_to_file/apriori.hst"\cr
(fichier(s) d'informations \emph{a priori}. Voir ci-dessous
pour la syntaxe. Exemple : hst_filename = "path_to_file/apriori.hst")}

\item{explanatory_vars}{Explanatory vars in a vector
Example : \code{c("CJ", "A21")} for the tabular \code{CJ} x \code{A21} \cr
( variables catégorielles, sous forme de vecteur.
Exemple : \code{c("CJ", "A21")} pour le premier
Pour un tableau croisant \code{CJ} x \code{A21})}

\item{value}{Colname for response variable in the tabular
\code{"<freq>"}. For frequency table \cr
(Nom de la variable de réponse dans le tableau
\code{"<freq>"}. Permet de tariter les tableaux de fréquence)}

\item{safety_rules}{Rules for primary suppression with Argus syntax, if the primary suppression
has been dealt with an apriori file specify manual safety range :"MAN(10)"
for example.\cr
( Règle(s) de secret primaire.
Chaîne de caractères en syntaxe batch Tau-Argus. Si le secret primaire
a été traité dans un fichier d'apriori : utiliser "MAN(10)")}

\item{suppress}{Algortihm for secondary suppression (Tau-Argus batch syntax), and the
parameters for it.\cr
( Algorithme de gestion du secret secondaire
(syntaxe batch de Tau-Argus), ainsi que les potentiels paramètres associés)}

\item{separator}{Character used as separator in the .tab file. \cr
(Caractère utilisé en tant que separateur dans le fichier .tab)}

\item{output_names}{output file name\cr
(nom du fichier en sortie.)}

\item{output_type}{Type of the output file (Argus codification)
By default \code{"2"} (csv for pivot-table).
For SBS files use \code{"4"}\cr
(Format des fichiers en sortie (codification Tau-Argus).
Valeur par défaut du package : \code{"2"} (csv for pivot-table).
Pour le format SBS utiliser \code{"4"})}

\item{output_options}{Additionnal parameter for the output,
by default : \code{"AS+"} (print Status). To specify no options : \code{""}.\cr
(Options supplémentaires des fichiers en sortie. Valeur
par défaut du package : \code{"AS+"} (affichage du statut). Pour ne
spécifier aucune option, \code{""}.)}

\item{gointeractive}{Boolean, if TRUE will open a Tau-Argus window and launch
the batch in it (\code{FALSE} by default). \cr
(Possibilité de lancer le batch depuis le menu de Tau-Argus
(\code{FALSE} par défaut).)}
}
\value{
A list containing two elements :
the arb file name and the output name (usefull if the name is generated randomly) \cr
(Une liste de deux éléments : le nom du fichier arb, le nom
fichiers en sortie (utile pour récupérer les noms générés aléatoirement).)
}
\description{
Function doesn't check if the asc and rda files exists\cr
(Crée un fichier batch (.arb) pour données tabulées exécutable par
Tau-Argus en ligne de commande. La fonction ne vérifie pas
si les fichiers asc et rda existent.)
}
\section{Syntaxe}{


For tabular data, this function secures only one table at a time#'
Unless specific details, use syntax from Tau-Argus manual \cr
Sauf mention contraire, utiliser la syntaxe mentionnée dans la documentation
de Tau-Argus.

Syntax pour \code{suppress} :
First parameter is the table number, in this case it will always be 1 \cr

(Pour les données tabulées la fonction traite un tableau à la fois
Syntaxe spéciale pour \code{suppress} :
le premier paramètre dans la syntaxe Tau-Argus est le numéro du tableau.
Dans le cas de cette fonction ce sera toujours 1)
}

\section{Informations \emph{hst_filename}}{

It's possible to add an apriori file (.hst) for a tabular, it can be generated
by the table_rda() function.

Other options are not mandatory. To modify default value use a vector with the
.hst file path as first element and then complete with those parameters :
\code{sep} for the separator,
\code{ignore_err} for IgnoreError and \code{exp_triv} for ExpandTrivial. \cr

(Il est possible de fournir un fichier apriori (.hst) pour un tableau,
il peut être fourni par la fonction table_rda()
Exemple : hst_filename = "mon_chemin/apriori.hst" \cr

Les options supplémentaires sont facultatives. Pour modifier les valeurs par
défaut, passer une liste ayant comme premier élément le(s) fichier(s) hst et
compléter avec les éléments portant les noms \code{sep} pour le séparateur,
\code{ignore_err} pour IgnoreError et \code{exp_triv} pour ExpandTrivial.
Comme pour les noms de fichiers, spécifier une seule valeur par paramètre ou
autant de valeurs que de tabulations.)
}

\examples{
\dontrun{
# creating arb file
infos_arb <- tab_arb(
  arb_filename = "path/tab1.arb",
  tab_filename = "path/tab1.tab",
  rda_filename = "path/tab1.rda",
  explanatory_vars = c("REGION", "CJ"),
  value = "CA",
  safety_rules = "NK(1,85)|FREQ(3,10)",
  suppress = "GH(1,100)",
  output_names = "path/tab1.csv",
  output_options = "AS+SE+",
  output_type = "4"
)
# show the content of the file in console
# visualisation du contenu du fichier dans la console
file.show(infos_arb$arb_filename, pager = "console")
}
}
