% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauargus_info.R
\name{tauargus_info}
\alias{tauargus_info}
\title{Informations sur Tau-Argus}
\usage{
tauargus_info()
}
\value{
Un vecteur caractère contenant trois composantes : path, version et
  date.
}
\description{
Recherche des informations sur la version de Tau-Argus utilisée. Cette
fonction est expérimentale.
}
\details{
La fonction lit le contenu d'un éventuel fichier TauNews.html (historique des
changements) et récupère les informations dans l'en-tête du dernier
paragraphe rédigé. Si le fichier n'est pas présent ou change de structure
d'une version sur l'autre, la fonction ne donnera pas le résultat attendu.
}
