% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tot5}
\alias{tot5}
\title{Total codes for the \code{dfs5_agg} dataset}
\format{
A named character vector (e.g., \code{c(ACT = "Total", GEO = "Total", ...)}).
}
\usage{
tot5
}
\description{
A named character vector specifying the code representing the Total for each
of the 5 variables in the \code{\link{dfs5_agg}} dataset (\code{ACT}, \code{GEO}, \code{SEX}, \code{AGE}, \code{ECO}).
}
\seealso{
\code{\link{dfs5_agg}}, \code{\link{hrc5}}
}
\keyword{datasets}
