% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turnover_act_cj}
\alias{turnover_act_cj}
\title{Turnover broken down by business sector and type of companies (fake values).}
\format{
A tibble/data frame with 406 rows and 5 variables:
\describe{
  \item{ACTIVITY}{business sector, hierarchical variables with three levels described
  in the activity_corr_table dataset. The root is noted "Total"}
  \item{CJ}{Type of companies (3 categories + overall category "Total")}
  \item{N_OBS}{Frequency, number of companies}
  \item{TOT}{turnover}
  \item{MAX}{turnover of the company which contributes the most to the cell.}
}
}
\usage{
turnover_act_cj
}
\description{
A tabular dataset containing the turnover broken down by Business sector
and Type of companies. Useful for playing with tab_ functions.
}
\seealso{
activity_corr_table
}
\keyword{datasets}
