% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turnover_act_size}
\alias{turnover_act_size}
\title{data - Turnover broken down by business sector and size of French companies (fake values).}
\format{
A tibble/data frame with 414 rows and 5 variables:
\describe{
 \item{ACTIVITY}{business sector, hierarchical variables with three levels described
  in the activity_corr_table dataset. The root is noted "Total"}
  \item{SIZE}{size of the companies (Number of employees in three categories
  + overall category "Total")}
  \item{N_OBS}{Frequency, number of companies}
  \item{TOT}{turnover value in euros}
  \item{MAX}{turnover of the company which contributes the most to the cell.}
}
}
\usage{
turnover_act_size
}
\description{
A tabular dataset containing the turnover broken down by Business sector
and Size of companies. Useful for playing with tab_ functions.
}
\seealso{
activity_corr_table
}
\keyword{datasets}
