% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datatest2}
\alias{datatest2}
\title{data crossing 5 categorical variables, none are hierarchical.}
\format{
A tibble/data frame with 5 612 rows and 15 variables:
\describe{
\item{A10}{business sector, not hierarchical}
\item{cj}{legal category, not hierarchical}
\item{type_distrib}{type of distribution, not hierarchical}
\item{treff}{Number of employees (categorical), not hierarchical}
\item{nuts1}{NUTS region, no hierarchical}
\item{nb_obs}{Frequency, number of companies}
\item{nb_obs_rnd}{Frequency rounded, number of companies}
\item{pizzas_tot}{turnover value in euros}
\item{pizzas_tot_abs}{turnover absolute value in euros}
\item{pizzas_max}{turnover max value in euros}
\item{is_secret_freq}{Boolean, TRUE if primary secret for frequency rule}
\item{is_secret_dom}{Boolean, TRUE if primary secret for dominance rule}
\item{is_secret_prim}{Boolean, TRUE if primary secret for any rule}

}
}
\usage{
datatest2
}
\description{
data crossing 5 categorical variables, none are hierarchical.
}
\keyword{datasets}
