% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Importe les résultats de Tau-Argus}
\usage{
import(arb_filename)
}
\arguments{
\item{arb_filename}{nom du fichier arb (avec extension) contenant les
informations nécessaires à l'import.}
}
\value{
Une liste d'un ou plusieurs data.frames. Chaque data.frame correspond
  au résultat d'une tabulation. Les noms des tableaux renseignés dans les
  lignes du batch de la forme \code{// <TABLE_ID> "..."} sont récupérés.
}
\description{
Importe dans R les résultats générés par Tau-Argus à partir des informations
contenues dans un fichier arb.
}
\details{
Nécessite que le batch ait été exécuté et se soit terminé sans erreur. Afin
d'importer immédiatement après exécution du batch, cette fonction sera ainsi
le plus souvent appelée via \code{\link{run_tauargus}} (en paramétrant
\code{import = TRUE}).

Il n'est possible (pour l'instant) que d'importer les résultats de type "2"
(csv for pivot-table) et "4" (sbs). En cas d'impossibilité de l'import pour
une tabulation donnée, un data.frame vide est retourné (avec un message
d'avertissement).
}
\section{Attributs}{


À chaque data.frame est associé un ensemble d'attributs (métadonnées)
permettant de conserver une trace des spécifications passées à Tau-Argus.

Attributs systématiquement présents :
\code{explanatory_vars}, \code{response_var}, \code{safetyrule},
\code{suppress}, \code{linked}, \code{output_type}.

Attributs présents uniquement si l'option correspondante a été renseignée par
l'utilisateur : \code{shadow_var}, \code{cost_var}, \code{output_options}.
}

