% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_arb.R
\name{micro_arb}
\alias{micro_arb}
\title{Crée un fichier batch (.arb) pour microdonnées}
\usage{
micro_arb(
  arb_filename = NULL,
  asc_filename,
  rda_filename = NULL,
  explanatory_vars,
  response_var = getOption("rtauargus.response_var"),
  shadow_var = NULL,
  cost_var = NULL,
  safety_rules,
  weighted = getOption("rtauargus.weighted"),
  suppress,
  linked = getOption("rtauargus.linked"),
  output_names = NULL,
  output_type = getOption("rtauargus.output_type"),
  output_options = getOption("rtauargus.output_options"),
  apriori = NULL,
  gointeractive = FALSE
)
}
\arguments{
\item{arb_filename}{nom du fichier arb généré (avec
extension). Si non renseigné, un fichier temporaire.}

\item{asc_filename}{[\strong{obligatoire}] nom du fichier asc (avec extension).}

\item{rda_filename}{nom du fichier rda (avec extension). Si non renseigné,
\code{asc_filename} avec l'extension "rda" à la place de "asc".}

\item{explanatory_vars}{[\strong{obligatoire}] variables catégorielles, sous
forme de liste de vecteurs. Chaque élément de la liste est un vecteur des
noms des variables formant une tabulation.
Exemple : \code{list(c("CJ", "A21"), c("SEXE", "REGION"))} pour le premier
tableau croisant \code{CJ} x \code{A21} et le deuxième tableau croisant
\code{SEXE} x \code{REGION}.
Si une seule tabulation, un simple vecteur des variables à croiser est
accepté (pas besoin de \code{list(...)}).}

\item{response_var}{variable de réponse à sommer, ou comptage si
\code{"<freq>"}. Une seule valeur ou autant de valeurs que de tabulations.}

\item{shadow_var}{variable(s) pour l'application du secret primaire. Si non
renseigné, \code{response_var} sera utilisé par Tau-Argus.}

\item{cost_var}{variable(s) de coût pour le secret secondaire.}

\item{safety_rules}{[\strong{obligatoire}] règle(s) de secret primaire.
Chaîne de caractères en syntaxe batch Tau-Argus. La pondération est traitée
dans un paramètre à part (ne pas spécifier WGT ici, utiliser le paramètre
\code{weighted}).}

\item{weighted}{indicatrice(s) de pondération (booléen).}

\item{suppress}{[\strong{obligatoire}] méthode(s) de gestion du secret
secondaire (syntaxe batch de Tau-Argus). Si la méthode est la même pour
chaque tabulation, le premier paramètre (numéro du tableau) sera ignoré et
renuméroté automatiquement (voir la section 'Syntaxe').}

\item{linked}{pour traiter le secret secondaire conjointement sur toutes les
tabulations. Une seule commande suppress autorisée dans ce cas (appliquée à
tous les tableaux).}

\item{output_names}{noms des fichiers en sortie. Si renseigné,
obligatoirement autant de noms de fichiers que de tabulations. Si laissé
vide, autant de noms de fichiers temporaires que de tabulations seront
générés.}

\item{output_type}{format des fichiers en sortie (codification Tau-Argus).
Valeur par défaut du package : \code{"2"} (csv for pivot-table).}

\item{output_options}{options supplémentaires des fichiers en sortie. Valeur
par défaut du package : \code{"AS+"} (affichage du statut). Pour ne
spécifier aucune option, \code{""}.}

\item{apriori}{fichier(s) d'informations \emph{a priori}. Voir ci-dessous
pour la syntaxe.}

\item{gointeractive}{pour avoir la possibilité de lancer le batch depuis le
menu de Tau-Argus (\code{FALSE} par défaut).}
}
\value{
Une liste de deux éléments : le nom du fichier arb, les noms des
  fichiers en sortie (utile pour récupérer les noms générés aléatoirement).
}
\description{
Crée un fichier batch pour microdonnées, exécutable par Tau-Argus en ligne de
commande.
}
\details{
La fonction ne vérifie pas si les fichiers asc et rda existent.
}
\section{Syntaxe}{

Tau-Argus peut traiter plusieurs tabulations pour un même jeu de
microdonnées. Passer une seule valeur pour une option appliquera le même
traitement à chaque tabulation. Pour des options différenciées, passer un
vecteur contenant autant de valeurs que de tabulations. Si les longueurs ne
correspondent pas (déconseillé), un recyclage est effectué.

Sauf mention contraire, utiliser la syntaxe mentionnée dans la documentation
de Tau-Argus.

Syntaxe spéciale pour \code{suppress} : le premier paramètre dans la
syntaxe Tau-Argus est le numéro de la tabulation. Si la méthode est identique
pour toutes les tabulations, ce premier paramètre sera ignoré et les numéros
recalculés automatiquement pour le batch. Dans l'écriture
\code{suppress = "GH(n,100)"}, n sera ainsi transformé en 1 pour la première
tabulation, en 2 pour la deuxième tabulation, etc.
}

\section{Identifiants des tableaux}{

Si la liste \code{explanatory_vars} comporte des noms, ceux-ci seront
utilisés dans le batch pour donner un identifiant au tableau, sous la forme
d'une ligne de commentaire (\code{// <TABLE_ID> "..."}). Ils seront
réutilisés par la fonction \code{import} pour nommer les tableaux formats R
en sortie.
}

\section{Informations \emph{a priori}}{

Il est possible de fournir un fichier a priori (.hst) pour chaque tabulation.

La manière la plus simple est de passer un vecteur contenant autant de noms de
fichiers hst que de tabulations. Si le fichier est le même pour toutes les
tabulations, spécifier le nom de ce fichier (il sera utilisé pour toutes les
tabulations).

Les options supplémentaires sont facultatives. Pour modifier les valeurs par
défaut, passer une liste ayant comme premier élément le(s) fichier(s) hst et
compléter avec les éléments portant les noms \code{sep} pour le séparateur,
\code{ignore_err} pour IgnoreError et \code{exp_triv} pour ExpandTrivial.
Comme pour les noms de fichiers, spécifier une seule valeur par paramètre ou
autant de valeurs que de tabulations.
}

\section{Voir aussi}{
 La fonction \code{\link{rtauargus}}, qui utilise cette
fonction et hérite de ses paramètres.
}

\examples{
# creation fichier arb
infos_arb <- micro_arb(
  asc_filename = "donnees.asc",
  explanatory_vars = list(c("REGION", "CJ"), c("REGION")),
  response_var = c("CA", "<freq>"),
  safety_rules = c("NK(1,85)|FREQ(3,10)", "FREQ(3,10)"),
  suppress = "GH(.,100)",
  output_names = c("tab1.csv", "~/tab2.csv"),
  output_options = c("AS+SE+", "SE+"),
  output_type = "2"
)

# visualisation du contenu du fichier dans la console
file.show(infos_arb$arb_filename, pager = "console")
}
