% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asc_rda.R
\name{micro_asc_rda}
\alias{micro_asc_rda}
\title{Crée les fichiers asc et rda à partir de microdonnées}
\usage{
micro_asc_rda(microdata, asc_filename = NULL, rda_filename = NULL,
  weight_var = NULL, holding_var = NULL, hrc = NULL,
  decimals = getOption("rtauargus.decimals"),
  totcode = getOption("rtauargus.totcode"),
  missing = getOption("rtauargus.missing"), codelist = NULL,
  request_var = NULL, request_code = NULL)
}
\arguments{
\item{microdata}{[\strong{obligatoire}] data.frame contenant les
microdonnées.}

\item{asc_filename}{nom du fichier asc (avec extension). Si non renseigné, un
fichier temporaire.}

\item{rda_filename}{nom du fichier rda (avec extension). Si non renseigné,
\code{asc_filename} avec l'extension "rda" à la place de "asc").}

\item{weight_var}{nom de la variable de poids.}

\item{holding_var}{nom de la variable de holding.}

\item{hrc}{informations des variables hiérarchiques. Une liste dont les noms
sont les variables concernées et les éléments sont les informations.
Exemple :
\code{list(}
\code{VAR1 = c(hierleadstring = "@", hiercodelist = "hrc/var1.hrc"),}
\code{VAR3 = c(hierlevels = "3 1 0")}
\code{)}}

\item{decimals}{nombre minimal de décimales à afficher (voir section
'Details').}

\item{totcode}{code(s) pour le total d'une variable catégorielle (voir
section 'Paramètres spécifiques' pour la syntaxe de ce paramètre). Les
variables non spécifiées (ni par défaut, ni explicitement) se verront
attribuer la valeur de \code{rtauargus.totcode}.}

\item{missing}{code(s) pour une valeur manquante (voir section
'Paramètres spécifiques' pour la syntaxe de ce paramètre).}

\item{codelist}{fichier(s) contenant les libellés des variables catégorielles
(voir section 'Paramètres spécifiques' pour la syntaxe de ce paramètre).}

\item{request_var}{(pas encore implémenté)}

\item{request_code}{(pas encore implémenté)}
}
\value{
Renvoie les noms des fichiers asc et rda sous forme de liste (de
  manière invisible)
}
\description{
Crée un fichier texte de longueur fixe (asc) et un fichier de métadonnées
(rda) à partir de microdonnées et d'informations additionnelles.
}
\details{
Le paramètre \code{decimals} indique le nombre minimal de décimales à faire
figurer dans le fichier en sortie (quel que soit le nombre de décimales
effectivement présent dans \code{microdata}). Il s'applique à toutes les
variables réelles (double) mais pas aux variables entières (integer). Par
exemple, pour ajouter des zéros à une variable entière, la convertir avec
\code{as.double} au préalable.

Les chiffres après la virgule peuvent être incorrects si le nombre total de
chiffres (avant ou après le séparateur décimal) est supérieur à 15. Voir
\code{\link[gdata]{write.fwf}} (fonction utilisée pour écrire le fichier asc)
pour plus de détails.

Les colonnes vides (\code{NA} ou chaînes de caractère vide) ne seront pas
exportées dans le fichier asc. Un message d'avertissement listera les
colonnes concernées.
}
\section{Paramètres spécifiques}{


Les paramètres \code{totcode}, \code{missing} et \code{codelist} sont
à renseigner sous la forme d'un vecteur indiquant la valeur à prendre pour
chaque variable.

Les noms des éléments du vecteur donnent la variable concernée, les éléments
du vecteur donnent la valeur du paramètre pour Tau-Argus. Un élément non
nommé constituera la valeur par défaut, qui sera attribuée à toutes les
variables pouvant prendre ce paramètre.

Exemples :
\itemize{
  \item{\code{totcode = "Ensemble"} écrira \code{<TOTCODE> "Ensemble"} pour
    toutes les variables (catégorielles)}
  \item{\code{totcode = c("Ensemble", GEO = "France")} fera une exception pour
   la variable \code{GEO}}
}
}

\examples{
# donnees fictives temporaires
micro_df <-
  data.frame(
    GEO   = c("443", "541", "543"),
    A21   = c("A", "B", "A"),
    CA    = c(100, 0, 7),
    POIDS = c(1, 2.71, 4.2)
  )

res <- micro_asc_rda(
  microdata  = micro_df,
  weight_var = "POIDS",
  decimals   = 1,
  hrc        = list(GEO = c(hierlevels = "2 1")),
  totcode    = c(GEO = "France")
)

# visualisation des fichiers produits
res # (noms des fichiers)
file.show(res$asc_filename, pager = "internal", title = "fichier asc")
file.show(res$rda_filename, pager = "internal", title = "fichier rda")
}
