% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asc_rda.R
\name{micro_asc_rda}
\alias{micro_asc_rda}
\title{Crée les fichiers asc et rda à partir de microdonnées}
\usage{
micro_asc_rda(microdata, asc_filename = NULL, rda_filename = NULL,
  weight = NULL, holding = NULL, hrc = NULL, decimals = 0,
  totcode = "Total", missing = NULL, codelist = NULL, request = NULL,
  request_code = NULL)
}
\arguments{
\item{microdata}{[\strong{obligatoire}] data.frame contenant les microdonnées.}

\item{asc_filename}{nom du fichier asc (avec extension). Si non renseigné, un
fichier temporaire.}

\item{rda_filename}{nom du fichier rda (avec extension). Si non renseigné,
\code{asc_filename} avec l'extension "rda" à la place de "asc").}

\item{weight}{nom de la variable de poids.}

\item{holding}{nom de la variable de holding.}

\item{hrc}{informations des variables hiérarchiques. Une liste dont les noms
sont les variables concernées et les éléments sont les informations.
Exemple :
\code{list(}
\code{VAR1 = c(hierleadstring = "@", hiercodelist = "hrc/var1.hrc"),}
\code{VAR3 = c(hierlevels = "3 1 0")}
\code{)}}

\item{decimals}{nombre de décimales minimum à afficher, par défaut 0 (cf.
section 'Details').}

\item{totcode}{(pas encore implémenté)}

\item{missing}{(pas encore implémenté)}

\item{codelist}{(pas encore implémenté)}

\item{request}{(pas encore implémenté)}

\item{request_code}{(pas encore implémenté)}
}
\value{
Renvoie les noms des fichiers asc et rda sous forme de liste (de
  manière invisible).
}
\description{
Crée un fichier texte de longueur fixe (asc) et un fichier de métadonnées
(rda) à partir de microdonnées et d'informations additionnelles.
}
\details{
Le paramètre \code{decimals} indique le nombre minimal de décimales à faire
figurer dans le fichier en sortie (quel que soit le nombre de décimales
effectivement présent dans \code{microdata}). Il s'applique à toutes les
variables réelles (double) mais pas aux variables entières (integer). Par
exemple, pour ajouter des zéros à une variable entière, la convertir avec
\code{as.double} au préalable.

Les chiffres après la virgule peuvent être incorrects si le nombre total de
chiffres (avant ou après le séparateur décimal) est supérieur à 15. Voir
\code{\link[gdata]{write.fwf}} (fonction utilisée pour écrire le fichier asc)
pour plus de détails.
}
\examples{
# donnees fictives temporaires
micro_df <- data.frame(
  ARRONDISSEMENT = c("443", "541", "543"),
              CA = c(100, 0, 7),
           POIDS = c(1, 2.71, 4.2)
)

res <- micro_asc_rda(
  micro_df,
  weight = "POIDS",
  decimals = 1,
  hrc = list(ARRONDISSEMENT = c(hierlevels = "2 1"))
)

# visualisation des fichiers produits
res # (noms des fichiers)
file.show(res$asc_filename, pager = "internal", title = "asc")
file.show(res$rda_filename, pager = "internal", title = "rda")
}
