% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_asc_rda.R
\name{micro_asc_rda}
\alias{micro_asc_rda}
\title{Crée les fichiers asc et rda à partir de microdonnées}
\usage{
micro_asc_rda(microdata, asc_filename = NULL, rda_filename = NULL,
  weight_var = NULL, holding_var = NULL,
  decimals = getOption("rtauargus.decimals"), hrc = NULL,
  hierleadstring = getOption("rtauargus.hierleadstring"),
  totcode = getOption("rtauargus.totcode"),
  missing = getOption("rtauargus.missing"), codelist = NULL,
  request = NULL)
}
\arguments{
\item{microdata}{[\strong{obligatoire}] data.frame contenant les
microdonnées.}

\item{asc_filename}{nom du fichier asc (avec extension). Si non renseigné, un
fichier temporaire.}

\item{rda_filename}{nom du fichier rda (avec extension). Si non renseigné,
\code{asc_filename} avec l'extension "rda" à la place de "asc".}

\item{weight_var}{nom de la variable de poids.}

\item{holding_var}{nom de la variable de holding.}

\item{decimals}{nombre minimal de décimales à afficher (voir section 'Nombre
de décimales').}

\item{hrc}{informations sur les variables hiérarchiques (voir section
'Variables hiérarchiques').}

\item{hierleadstring}{caractère qui, répété n fois, indique que la valeur est
à n niveaux de profondeur dans la hiérarchie.}

\item{totcode}{code(s) pour le total d'une variable catégorielle (voir
section 'Paramètres spécifiques' pour la syntaxe de ce paramètre). Les
variables non spécifiées (ni par défaut, ni explicitement) se verront
attribuer la valeur de \code{rtauargus.totcode}.}

\item{missing}{code(s) pour une valeur manquante (voir section
'Paramètres spécifiques' pour la syntaxe de ce paramètre).}

\item{codelist}{fichier(s) contenant les libellés des variables catégorielles
(voir section 'Paramètres spécifiques' pour la syntaxe de ce paramètre).}

\item{request}{(pas encore implémenté)}
}
\value{
Renvoie les noms des fichiers asc et rda sous forme de liste (de
  manière invisible). Les colonnes vides (remplies de \code{NA} ou de chaînes
  de caractères vides) ne seront pas exportées dans le fichier asc. Un
  message d'avertissement listera les colonnes concernées.
}
\description{
Crée un fichier texte de longueur fixe (asc) et un fichier de métadonnées
(rda) à partir de microdonnées et d'informations additionnelles.
}
\section{Paramètres spécifiques}{


Les paramètres \code{hrc}, \code{totcode}, \code{missing} et \code{codelist}
sont à renseigner sous la forme d'un vecteur indiquant la valeur à prendre
pour chaque variable.

Les noms des éléments du vecteur donnent la variable concernée, les éléments
du vecteur donnent la valeur du paramètre pour Tau-Argus. Un élément non
nommé constituera la valeur par défaut, qui sera attribuée à toutes les
variables pouvant prendre ce paramètre.

Par exemple :
\itemize{
  \item{\code{totcode = "Ensemble"} : écrit \code{<TOTCODE> "Ensemble"} pour
    toutes les variables catégorielles}
  \item{\code{totcode = c("Ensemble", GEO = "France")} : idem, sauf pour la
    variable \code{GEO}}
}
}

\section{Variables hiérarchiques}{


Le paramètre \code{hrc} obéit aux mêmes règles de syntaxe que \code{totcode},
\code{missing} et \code{codelist} (éventuelle valeur commune, valeurs
spécifiques à chaque variable). Il présente de plus la particularité
d'accepter plusieurs façons de spécifier les valeurs associées aux variables
hiérarchiques.

Pour définir une hiérarchie basée sur les positions des caractères
(\strong{hierlevels}), passer une suite de nombre entiers séparés par des
espaces.

\emph{Exemple :} \code{c(CODECOM = "2 3 0 0 0")}

Si la hiérarchie est définie dans un fichier hrc à part
(\strong{hiercodelist}), la fonction attend l'emplacement de ce fichier (et
un éventuel \code{hierleadstring} s'il diffère de l'option par défaut du
package). Dans ce cas, on peut écrire explicitement le chemin vers un fichier
existant (\code{c(A38 = "a38.hrc")}), mais aussi aussi passer un appel à
\code{\link{write_hrc}} qui génèrera un fichier hrc à partir de microdonnées.

\emph{Exemple :} \code{c(A38 = write_hrc(microdata, c("A38", "A21", "A10")))}

Un raccourci pour cet appel est d'écrire les variables consituant la
hiérarchie séparées par des ">". Dans ce cas, les microdonnées et
hierleadstring qu'utilise \code{write_hrc} sont ceux déclarés dans
\code{micro_asc_rda}.

\emph{Exemple :} \code{c(A38 = "A38 > A21 > A10")} \emph{(nombre d'espaces
quelconque avant et après les ">")}

Les deux dernières méthodes passent par la création d'un fichier temporaire.
Pour un fichier hrc réutilisable, il est nécessaire de le créer au préalable
à l'aide de \code{write_hrc}.

Les trois méthodes nécessitent que les éléments du vecteur en paramètre
soient nommés (avec le nom de la variable), même s'il n'y a qu'un seul
élément.
}

\section{Nombre de décimales}{


Le paramètre \code{decimals} indique le nombre minimal de décimales à faire
figurer dans le fichier en sortie (quel que soit le nombre de décimales
effectivement présent dans \code{microdata}). Il s'applique à toutes les
variables réelles (double) mais pas aux variables entières (integer). Pour
ajouter des zéros à une variable entière, la convertir avec \code{as.double}
au préalable.

Les chiffres après la virgule peuvent être incorrects dans le fichier asc si
le nombre total de chiffres (avant ou après le séparateur décimal) est
supérieur à 15. Voir \code{\link[gdata]{write.fwf}} (fonction utilisée pour
écrire le fichier asc) pour plus de détails.
}

\section{Voir aussi}{
 La fonction \code{\link{rtauargus}}, qui utilise cette
fonction.
}

\examples{
# donnees fictives
micro_df <-
  data.frame(
    GEO   = c("443", "541", "543"),
    A10   = c( "AZ",  "BE",  "BE"),
    A21   = c(  "A",   "B",   "C"),
    DIPL  = c( "01",  "??",  "02"),
    CA    = c(  100,     0,     7),
    POIDS = c(    1,  2.71,   4.2)
  )

# creation asc + rda
res <-
  micro_asc_rda(
    microdata  = micro_df,
    weight_var = "POIDS",
    decimals   = 1,
    hrc        = c(A21 = "A21>A10", GEO = "2 1 0", DIPL = "dipl.hrc"),
    totcode    = c(GEO = "France"),
    missing    = c(DIPL = "??")
  )

# visualisation des fichiers produits
file.show(
  res$asc_filename, res$rda_filename,
  header = unlist(res),
  pager = "internal"
)
}
