% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtauargus-package.R
\docType{package}
\name{rtauargus-package}
\alias{rtauargus}
\alias{rtauargus-package}
\title{rtauargus: Using Tau-Argus from R}
\description{
Protects tables by calling the Tau-Argus software from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://inseefrlab.github.io/rtauargus}
  \item \url{https://github.com/inseefrlab/rtauargus}
  \item \url{https://inseefrlab.github.io/rtauargus/}
  \item Report bugs at \url{https://github.com/inseefrlab/rtauargus/issues}
}

}
\author{
\strong{Maintainer}: Julien Jamme \email{julien.jamme@insee.fr}

Authors:
\itemize{
  \item Pierre-Yves Berrard \email{pierre-yves.berrard@insee.fr}
  \item Nathanaël Rastout \email{nathanael.rastout@insee.fr}
  \item Jeanne Pointet
  \item Félix Beroud
  \item André-Raymond Socard
  \item Wistan Pomel
  \item Clara Baudry \email{clara.baudry@insee.fr}
}

Other contributors:
\itemize{
  \item  Institut National de la Statistique et des Études Économiques [copyright holder]
}

}
\keyword{internal}
