% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rtauargus_options}
\alias{rtauargus_options}
\alias{reset_rtauargus_options}
\title{Options du package rtauargus}
\usage{
rtauargus_options()

# options(rtauargus.<opt> = <val>)

reset_rtauargus_options(...)
}
\arguments{
\item{...}{noms des options à réinitialiser, séparés par des virgules. Si
aucun nom n'est spécifié, toutes les options du package seront
réinitialisées. Le préfixe \code{"rtauargus."} est facultatif.}
}
\description{
Gère les options du package (affiche, modifie, réinitialise).
}
\details{
Les options du package définissent les comportements par défaut des fonctions.

Ces options sont utilisées si un argument obligatoire d’une fonction n’est
pas renseigné. Le nom de l’option est le nom de l’argument d’une fonction
précédé de \code{rtauargus.} .

\emph{Par exemple, \code{rtauargus.decimals} sera la valeur utilisée si l’argument
\code{decimals} de la fonction \code{micro_asc_rda} n’est pas renseigné par
l’utilisateur.}

Au chargement, le package attribue une valeur par défaut à toutes les options
de rtauargus qui ne sont pas encore déclarées (cf. tableau ci-dessous). Les
options déjà définies par l'utilisateur gardent leurs valeurs.

Elles peuvent être redéfinies pour une session par une instruction
\code{options(rtauargus.}...\code{ = }...\code{)}, ou de manière globale si
de telles instructions sont placées dans un fichier de configuration propre à
l'utilisateur (fortement déconseillé si le programme a vocation à être
reproductible).

En cas d'effacement accidentel d'une option par l'utilisateur, les fonctions
utiliseront les valeurs par défaut du package.
}
\section{Liste des options}{

\tabular{lll}{
 \strong{Option}                  \tab  \strong{Valeur par défaut}                                    \tab \strong{Fonction}   \cr
 \code{------------------------}  \tab  \code{---------------------------------}                      \tab \code{-------------}\cr
 rtauargus.decimals               \tab  \Sexpr{rtauargus:::op.rtauargus$rtauargus.decimals}           \tab \link{micro_asc_rda}\cr
 rtauargus.totcode                \tab "\Sexpr{rtauargus:::op.rtauargus$rtauargus.totcode}"           \tab                     \cr
 rtauargus.missing                \tab "\Sexpr{rtauargus:::op.rtauargus$rtauargus.missing}"           \tab                     \cr
 rtauargus.hierleadstring         \tab "\Sexpr{rtauargus:::op.rtauargus$rtauargus.hierleadstring}"    \tab                     \cr
 \code{------------------------}  \tab  \code{---------------------------------}                      \tab \code{-------------}\cr
 rtauargus.response_var           \tab "\Sexpr{rtauargus:::op.rtauargus$rtauargus.response_var}"      \tab \link{micro_arb}    \cr
 rtauargus.weighted               \tab  \Sexpr{rtauargus:::op.rtauargus$rtauargus.weighted}           \tab                     \cr
 rtauargus.linked                 \tab  \Sexpr{rtauargus:::op.rtauargus$rtauargus.linked}             \tab                     \cr
 rtauargus.output_type            \tab "\Sexpr{rtauargus:::op.rtauargus$rtauargus.output_type}"       \tab                     \cr
 rtauargus.output_options         \tab "\Sexpr{rtauargus:::op.rtauargus$rtauargus.output_options}"    \tab                     \cr
 \code{------------------------}  \tab  \code{---------------------------------}                      \tab \code{-------------}\cr
 rtauargus.tauargus_exe           \tab "\Sexpr{rtauargus:::op.rtauargus$rtauargus.tauargus_exe}"      \tab \link{run_arb} \cr
 rtauargus.show_batch_console     \tab  \Sexpr{rtauargus:::op.rtauargus$rtauargus.show_batch_console} \tab                     \cr
 rtauargus.import                 \tab  \Sexpr{rtauargus:::op.rtauargus$rtauargus.import}             \tab
}
}

\examples{
rtauargus_options()

# modifie certaines options
options(
  rtauargus.tauargus_exe = "Z:/tmp/TauArgus.exe",
  rtauargus.output_type = "4",
  rtauargus.weighted = TRUE
)
str(rtauargus_options())

# reinitialise une partie des options (prefixe "rtauargus." facultatif)
reset_rtauargus_options("output_type", "rtauargus.tauargus_exe")
str(rtauargus_options())

# reinitialise tout
reset_rtauargus_options()
str(rtauargus_options())
}
\seealso{
\link{options}, le système d'options de R dans lequel s'insèrent les
  options de ce package.
}
