% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtauargus_plus.R
\name{rtauargus_plus}
\alias{rtauargus_plus}
\title{Secrétisation en masse}
\usage{
rtauargus_plus(
  grp_size = 5,
  microdata,
  explanatory_vars,
  safety_rules,
  suppress,
  ...
)
}
\arguments{
\item{grp_size}{nombre de tableaux par appel de Tau-Argus (un entier compris
entre 1 et 10).}

\item{microdata}{[\strong{obligatoire}] data.frame contenant les
microdonnées.}

\item{explanatory_vars}{[\strong{obligatoire}] variables catégorielles, sous
forme de liste de vecteurs. Chaque élément de la liste est un vecteur des
noms des variables formant une tabulation.
Exemple : \code{list(c("CJ", "A21"), c("SEXE", "REGION"))} pour le premier
tableau croisant \code{CJ} x \code{A21} et le deuxième tableau croisant
\code{SEXE} x \code{REGION}.
Si une seule tabulation, un simple vecteur des variables à croiser est
accepté (pas besoin de \code{list(...)}).}

\item{safety_rules}{[\strong{obligatoire}] règle(s) de secret primaire.
Chaîne de caractères en syntaxe batch Tau-Argus. La pondération est traitée
dans un paramètre à part (ne pas spécifier WGT ici, utiliser le paramètre
\code{weighted}).}

\item{suppress}{[\strong{obligatoire}] méthode de gestion du secret
secondaire (syntaxe batch de Tau-Argus). Une seule méthode autorisée pour
tous les tableaux. Exemple \code{"GH(.,100)"} (le point jouant le rôle du
numéro de tabulation).}

\item{...}{paramètres optionnels pour \code{micro_asc_rda}, \code{micro_arb}
et \code{run_arb}. Voir l'aide de ces fonctions.}
}
\value{
Une liste de data.frames (tableaux secrétisés).
}
\description{
Optimisation de la fonction \code{\link{rtauargus}} pour un grand nombre de
croisements (ayant tous les mêmes paramètres).
}
\details{
En mode interactif, Tau-Argus peut traiter jusqu'à 10 tabulations
simultanément pour un même jeu de microdonnées. En mode batch, un nombre plus
important de croisements peut être spécifié. Cependant, procéder de la sorte
peut être particulièrement long, car Tau-Argus prend beaucoup de temps
pour lire des fichiers texte volumineux de microdonnées.

\code{rtauargus_plus} permet d'améliorer la vitesse d'exécution. La fonction
découpe la liste des tabulations en groupes de taille \code{grp_size} et
effectue un appel à \code{rtauargus} pour chaque groupe. Elle écrit un
fichier asc restreint aux seules variables effectivement utilisées au sein
d'un groupe.

Les résultats sont ensuite agrégés en une unique liste, comme si Tau-Argus
n'avait été appelé qu'une seule fois.

Modifier \code{grp_size} ne changera pas le résultat, seulement le temps
d'exécution. Une valeur autour de 5 (défaut) semble être un bon compromis
entre une lecture de fichiers asc trop volumineux et un nombre d'appels à
Tau-Argus trop important. Elle peut être ajustée en fonction du nombre de
variables communes à l'intérieur de chaque groupe de tabulations.
}
\section{Limites par rapport à la fonction \code{rtauargus}}{


En contrepartie de la vitesse d'exécution, les croisements doivent avoir les
mêmes caractéristiques (mêmes règles de secret primaire, même méthode de
secret secondaire, même variable de pondération, etc.). Les paramètres
\code{safety_rules}, \code{supress}, ..., doivent donc contenir une valeur
unique.

De plus, il n'est pas possible de spécifier \code{asc_filename},
\code{rda_filename}, \code{arb_filename} ou \code{output_names} pour
récupérer les fichiers intermédiaires. Ces fichiers seront écrits dans un
dossier temporaire (et écrasés à chaque nouveau groupe). Par conséquent,
spécifier \code{import = FALSE} est sans objet et sera ignoré.

Les données doivent obligatoirement être un data.frame (fichiers asc et rda
pas autorisés).

Si l'option \code{linked} est utilisée, la liaison ne sera effective qu'à
l'intérieur de chaque groupe de tabulations.
}

\examples{
\dontrun{
# exemple de ?rtauargus, avec une repetition (artificielle) des croisements
rtauargus_plus(
  grp_size         = 6, # (pour lancer les 12 croisements en 2 appels)
  microdata        = data.frame(V1 = c("A", "A", "B", "C", "A"), V2 = "Z"),
  explanatory_vars = rep(list("V1", "V2", c("V1", "V2")), times = 4),
  safety_rules     = "FREQ(3,10)",
  suppress         = "GH(.,100)"
)}
}
\seealso{
\code{\link{rtauargus}}, fonction appelée de manière répétée par
  \code{rtauargus_plus}.
}
