% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tauargus.R
\name{run_tauargus}
\alias{run_tauargus}
\title{Exécute un batch Tau Argus}
\usage{
run_tauargus(arb_filename, logbook = NULL, ...)
}
\arguments{
\item{arb_filename}{nom du fichier batch à exécuter.}

\item{logbook}{nom du fichier de log (optionnel).}

\item{...}{paramètres supplémentaires pour \code{system()}. Par exemple,
\code{show.output.on.console} = \code{FALSE}, pour ne pas afficher la log
dans la console.}
}
\description{
Exécute les instructions contenues dans un fichier .arb pour Tau Argus.
}
\details{
L'emplacement du programme TauArgus.exe est défini de manière globale au
chargement du package et modifiable par \code{\link{set_tauargus_exe}}.
}
