% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tauargus.R
\name{run_tauargus}
\alias{run_tauargus}
\title{Exécute un batch Tau-Argus}
\usage{
run_tauargus(arb_filename,
  tauargus_exe = getOption("rtauargus.tauargus_exe"), logbook = NULL,
  show_batch_console = getOption("rtauargus.show_batch_console"),
  import = getOption("rtauargus.import"), ...)
}
\arguments{
\item{arb_filename}{nom du fichier batch à exécuter.}

\item{tauargus_exe}{répertoire et nom du logiciel Tau-Argus.}

\item{logbook}{nom du fichier où est enregistré le journal d'erreurs
(optionnel).}

\item{show_batch_console}{pour afficher le déroulement du batch dans la
console.}

\item{import}{pour importer dans R les fichiers produits, \code{TRUE} par
défaut.}

\item{...}{paramètres supplémentaires pour \code{system()}.}
}
\value{
\itemize{
  \item{une liste de data.frame contenant les résultats si
    \code{import = TRUE} (via la fonction \code{\link{import}})} ;
  \item{\code{NULL} sinon}.
}
}
\description{
Exécute les instructions contenues dans un fichier .arb pour Tau-Argus.
}
\details{
L'emplacement du programme TauArgus.exe est défini de manière globale au
chargement du package. De fait, l'argument \code{tauargus_exe} n'aura
normalement pas à être spécifié (sauf pour surcharger l'option globale le
temps de l'exécution de la foncion).
}
