% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitable.R
\name{tab_multi_manager}
\alias{tab_multi_manager}
\title{Manages the secondary secret of a list of tables}
\usage{
tab_multi_manager(
  list_tables,
  list_explanatory_vars,
  dir_name = NULL,
  hrc = NULL,
  totcode = getOption("rtauargus.totcode"),
  value = "value",
  freq = "freq",
  secret_var = "is_secret_prim",
  cost_var = NULL,
  func_to_call = "tab_rtauargus2",
  ip_start = 10,
  ip_end = 0,
  num_iter_max = 1000,
  ...
)
}
\arguments{
\item{list_tables}{named list of dataframes representing the tables to protect}

\item{list_explanatory_vars}{named list of character vectors of explanatory variables of each table mentionned in list_tables. Names of the list are the same as of the list of tables.}

\item{dir_name}{character, directory where to save tauargus files}

\item{hrc}{named character vector, indicated the .hrc file's path of each hierarchical variables. The names of the vector are the names of the corresponding explanatory variable.}

\item{totcode}{character vector: either a one length vector if all explanatory variable has the same total code (default to "Total") or a named character vector detailing the total code for each explanatory variable.}

\item{value}{character : name of the response variable in the tables (mandatory)}

\item{freq}{character: name of the frequency variable in the tables (mandatory)}

\item{secret_var}{character: name of the boolean variable indicating if the cell doesn't respect the primary rules (apriori) - mandatory}

\item{cost_var}{character: name of the cost variable (default to NULL)}

\item{func_to_call}{character: name of the function to use to apply the secret (calling tau-argus). By default, the function is \code{tab_rtauargus2}}

\item{ip_start}{integer: Interval protection level to apply at first treatment of each table}

\item{ip_end}{integer: Interval protection level to apply at other treatments}

\item{num_iter_max}{integer: Maximum of treatments to do on each table}

\item{...}{other arguments of func_to_call}
}
\value{
original list of tables. Secret Results of each iteration is added to each table.
For example, the result of first iteration is called 'is_secret_1' in each table.
It's a boolean variable, whether the cell has to be masked or not.
}
\description{
Manages the secondary secret of a list of tables
}
\examples{
library(rtauargus)
library(dplyr)
data(turnover_act_size)
data(turnover_act_cj)
data(activity_corr_table)

#0-Making hrc file of business sectors ----
hrc_file_activity <- activity_corr_table \%>\%
  write_hrc2(file_name = "hrc/activity")

#1-Prepare data ----
#Indicate whether each cell complies with the primary rules
#Boolean variable created is TRUE if the cell doesn't comply.
#Here the frequency rule is freq in (0;3)
#and the dominance rule is NK(1,85)
list_data_2_tabs <- list(
  act_size = turnover_act_size,
  act_cj = turnover_act_cj
) \%>\%
purrr::map(
  function(df){
    df \%>\%
      mutate(
        is_secret_freq = N_OBS > 0 & N_OBS < 3,
        is_secret_dom = ifelse(MAX == 0, FALSE, MAX/TOT>0.85),
        is_secret_prim = is_secret_freq | is_secret_dom
      )
  }
)
\dontrun{
options(
  rtauargus.tauargus_exe =
    "Y:/Logiciels/TauArgus/TauArgus4.2.2b1/TauArgus.exe"
)
res_1 <- tab_multi_manager(
  list_tables = list_data_2_tabs,
  list_explanatory_vars = list(
    act_size = c("ACTIVITY", "SIZE"),
    act_cj = c("ACTIVITY", "CJ")
  ),
  hrc = c(ACTIVITY = hrc_file_activity),
  dir_name = "tauargus_files",
  value = "TOT",
  freq = "N_OBS",
  secret_var = "is_secret_prim",
  totcode =  "Total"
)
}

}
\seealso{
\code{tab_rtauargus2}
}
