% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitable.R
\name{tab_multi_manager}
\alias{tab_multi_manager}
\title{Manages the secondary secret of a list of tables}
\usage{
tab_multi_manager(
  list_tables,
  list_explanatory_vars,
  dir_name = NULL,
  hrc = NULL,
  alt_hrc = NULL,
  totcode = getOption("rtauargus.totcode"),
  alt_totcode = NULL,
  value = "value",
  freq = "freq",
  secret_var = "is_secret_prim",
  cost_var = NULL,
  suppress = "MOD(1,5,1,0,0)",
  ip_start = 10,
  ip_end = 0,
  num_iter_max = 10,
  ...
)
}
\arguments{
\item{list_tables}{named list of dataframes representing the tables to protect}

\item{list_explanatory_vars}{named list of character vectors of explanatory
variables of each table mentionned in list_tables. Names of the list are the same as of the list of tables.}

\item{dir_name}{string indicated the path of the directory in which to save
all the files (.rda, .hst, .txt, .arb, .csv) generated by the function.}

\item{hrc}{Informations of hierarchical variables (see section
'Hierarchical variables'). \cr
(Informations sur les variables hiérarchiques (voir section
'Hierarchical variables').)}

\item{alt_hrc}{named list for alternative hierarchies (useful for non nested-hierarchies)}

\item{totcode}{Code(s) which represent the total of a categorical variable
(see section 'Specific parameters' for this parameter's syntax).
If unspecified for a variable(neither by default nor explicitly)
it will be set to  \code{rtauargus.totcode}. \cr
(Code(s) pour le total d'une variable catégorielle (voir
section 'Specific parameters' pour la syntaxe de ce paramètre). Les
variables non spécifiées (ni par défaut, ni explicitement) se verront
attribuer la valeur de \code{rtauargus.totcode}.)}

\item{alt_totcode}{named list for alternative codes}

\item{value}{Name of the column containing the value of the cells. \cr
(Nom de la colonne contenant la valeur des cellules)}

\item{freq}{Name of the column containing the cell frequency. \cr
(Nom de la colonne contenant les effectifs pour une cellule)}

\item{secret_var}{Boolean variable which give the primary secret : equal to
"TRUE" if a cell is concerned by the primary secret,"FALSE" otherwise.
will  be exported in the apriori file \cr
(Variable indiquant le secret primaire de type booléen:
prend la valeur "TRUE" quand les cellules du tableau doivent être masquées
par le secret primaire, "FALSE" sinon. Permet de créer un fichier d'apriori)}

\item{cost_var}{Numeric variable allow to change the cost suppression of a cell
for secondary suppression, it's the value of the cell by default, can be
specified for each cell, fill with NA if the cost doesn't need to be changed
for all cells \cr
(Variable numeric qui permet de changer la coût de suppression d'une cellule,
pris en compte dans les algorithmes de secret secondaire.Par défaut le coût
correspond à la valeur de la cellule.  peut être spécifié pour chacune des cellules,
peut contenir des NA pour les coûts que l'on ne souhaite pas modifier.)}

\item{suppress}{[\strong{mandatory}]
Algortihm for secondary suppression (Tau-Argus batch syntax), and the
parameters for it.\cr
([\strong{obligatoire}] Algorithme de gestion du secret secondaire
(syntaxe batch de Tau-Argus), ainsi que les potentiels paramètres associés)}

\item{ip_start}{integer: Interval protection level to apply at first treatment of each table}

\item{ip_end}{integer: Interval protection level to apply at other treatments}

\item{num_iter_max}{integer: Maximum of treatments to do on each table (default to 10)}

\item{...}{other arguments of \code{tab_rtauargus2()}}
}
\value{
original list of tables. Secret Results of each iteration is added to each table.
For example, the result of first iteration is called 'is_secret_1' in each table.
It's a boolean variable, whether the cell has to be masked or not.
}
\description{
Manages the secondary secret of a list of tables
}
\examples{
library(rtauargus)
library(dplyr)
data(turnover_act_size)
data(turnover_act_cj)
data(activity_corr_table)

#0-Making hrc file of business sectors ----
hrc_file_activity <- activity_corr_table \%>\%
  write_hrc2(file_name = "hrc/activity")

#1-Prepare data ----
#Indicate whether each cell complies with the primary rules
#Boolean variable created is TRUE if the cell doesn't comply.
#Here the frequency rule is freq in (0;3)
#and the dominance rule is NK(1,85)
list_data_2_tabs <- list(
  act_size = turnover_act_size,
  act_cj = turnover_act_cj
) \%>\%
purrr::map(
  function(df){
    df \%>\%
      mutate(
        is_secret_freq = N_OBS > 0 & N_OBS < 3,
        is_secret_dom = ifelse(MAX == 0, FALSE, MAX/TOT>0.85),
        is_secret_prim = is_secret_freq | is_secret_dom
      )
  }
)
\dontrun{
options(
  rtauargus.tauargus_exe =
    "Y:/Logiciels/TauArgus/TauArgus4.2.2b1/TauArgus.exe"
)
res_1 <- tab_multi_manager(
  list_tables = list_data_2_tabs,
  list_explanatory_vars = list(
    act_size = c("ACTIVITY", "SIZE"),
    act_cj = c("ACTIVITY", "CJ")
  ),
  hrc = c(ACTIVITY = hrc_file_activity),
  dir_name = "tauargus_files",
  value = "TOT",
  freq = "N_OBS",
  secret_var = "is_secret_prim",
  totcode =  "Total"
)
}

}
\seealso{
\code{tab_rtauargus2}
}
