% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_rda.R
\name{tab_rda}
\alias{tab_rda}
\title{Crée les fichiers rda à partir de données tabulées}
\usage{
tab_rda(
  tabular,
  tab_filename = NULL,
  rda_filename = NULL,
  decimals = getOption("rtauargus.decimals"),
  hrc = NULL,
  hierleadstring = getOption("rtauargus.hierleadstring"),
  totcode = getOption("rtauargus.totcode"),
  missing = getOption("rtauargus.missing"),
  codelist = NULL,
  value = NULL,
  freq = NULL,
  maxscore = NULL,
  separator = ","
)
}
\arguments{
\item{tabular}{[\strong{obligatoire}] data.frame contenant les
données tabulées}

\item{tab_filename}{nom du fichier tab (avec extension)}

\item{rda_filename}{nom du fichier rda (avec extension).}

\item{decimals}{nombre minimal de décimales à afficher (voir section 'Nombre
de décimales').}

\item{hrc}{informations sur les variables hiérarchiques (voir section
'Variables hiérarchiques').}

\item{hierleadstring}{caractère qui, répété n fois, indique que la valeur est
à n niveaux de profondeur dans la hiérarchie.}

\item{totcode}{code(s) pour le total d'une variable catégorielle (voir
section 'Paramètres spécifiques' pour la syntaxe de ce paramètre). Les
variables non spécifiées (ni par défaut, ni explicitement) se verront
attribuer la valeur de \code{rtauargus.totcode}.}

\item{missing}{code(s) pour une valeur manquante (voir section
'Paramètres spécifiques' pour la syntaxe de ce paramètre).}

\item{codelist}{fichier(s) contenant les libellés des variables catégorielles
(voir section 'Paramètres spécifiques' pour la syntaxe de ce paramètre).}

\item{value}{nom de la colonne contenant la valeur des cellules}

\item{freq}{nom de la colonne contenant les effectifs pour une cellule}

\item{maxscore}{nom de la colonne contenant la valeur du plus gros contributeur
d'une cellule}

\item{separator}{charactere utilisé en tant que separateur dans le fichier .tab}
}
\value{
Renvoie le nom du fichier rda sous forme de liste (de
  manière invisible).
}
\description{
Crée un fichier tabular (tab) et de métadonnées
(rda) à partir de données tabulées et d'informations additionnelles.
}
\section{Paramètres spécifiques}{


Les paramètres \code{totcode}, \code{missing} et \code{codelist}
sont à renseigner sous la forme d'un vecteur indiquant la valeur à prendre
pour chaque variable.

Les noms des éléments du vecteur donnent la variable concernée, les éléments
du vecteur donnent la valeur du paramètre pour Tau-Argus. Un élément non
nommé constituera la valeur par défaut, qui sera attribuée à toutes les
variables pouvant prendre ce paramètre.

Par exemple :
\itemize{
  \item{\code{totcode = "Ensemble"} : écrit \code{<TOTCODE> "Ensemble"} pour
    toutes les variables catégorielles}
  \item{\code{totcode = c("Ensemble", GEO = "France")} : idem, sauf pour la
    variable \code{GEO}}
}
}

\section{Variables hiérarchiques}{


Le paramètre \code{hrc} obéit aux mêmes règles de syntaxe que \code{totcode},
\code{missing} et \code{codelist} (vecteur nommé contenant autant d'éléments
que de variables à décrire). Il présente de plus la particularité
d'accepter plusieurs façons de spécifier les valeurs associées aux variables
hiérarchiques.

Pour définir une hiérarchie basée sur les positions des caractères
(\strong{hierlevels}), passer une suite de nombre entiers séparés par des
espaces.

\emph{Exemple :} \code{c(CODECOM = "2 3 0 0 0")}

Si la hiérarchie est définie dans un fichier hrc à part
(\strong{hiercodelist}), la fonction attend l'emplacement de ce fichier (et
un éventuel \code{hierleadstring} s'il diffère de l'option par défaut du
package). Dans ce cas, on peut écrire explicitement le chemin vers un fichier
existant (\code{c(A38 = "a38.hrc")}), mais aussi passer un appel à
\code{\link{write_hrc}} qui génèrera un fichier hrc à partir de données tabulées.

\emph{Exemple :} \code{c(A38 = write_hrc(tabular, c("A38", "A21", "A10")))}


Les deux dernières méthodes passent par la création d'un fichier temporaire.
Pour un fichier hrc réutilisable, il est nécessaire de le créer au préalable
à l'aide de \code{write_hrc}.

Les trois méthodes nécessitent que les éléments du vecteur en paramètre
soient nommés (avec le nom de la variable), même s'il n'y a qu'un seul
élément.
}

\section{Nombre de décimales}{


Le paramètre \code{decimals} indique le nombre minimal de décimales à faire
figurer dans le fichier en sortie (quel que soit le nombre de décimales
effectivement présent dans \code{tabular}). Il s'applique à toutes les
variables réelles (double) mais pas aux variables entières (integer). Pour
ajouter des zéros à une variable entière, la convertir avec \code{as.double}
au préalable.
}

\section{Voir aussi}{
 La fonction \code{\link{rtauargus}}, qui utilise cette
fonction et hérite de ses paramètres.
}

\examples{
\dontrun{
# donnees fictives
tab_df <-
  data.frame(
    GEO      = c("443", "541", "543"),
    A10      = c( "AZ",  "BE",  "BE"),
    A21      = c(  "A",   "B",   "C"),
    DIPL     = c( "01",  "??",  "02"),
    CA       = c(  100,     0,     7),
    effectif = c(    3,     6,     8),
    max_CA   = c(   57,     0,     1),
  )

# creation rda
res <-
  tab_rda(
    tabular  = tab_df,
    decimals   = 1,
    hrc        = c(A21 = "A21>A10", GEO = "2 1 0", DIPL = "dipl.hrc"),
    totcode    = c(GEO = "France"),
    missing    = c(DIPL = "??"),
    value = "CA",
    freq = "effectif",
    maxscore = "max_CA",
    separator = ","
  )

# visualisation des fichiers produits
file.show(
 res$rda_filename,
  header = unlist(res),
  pager = "internal"
)
}
}
