% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_rda.R
\name{tab_rda}
\alias{tab_rda}
\title{Creates rda files from tabular data}
\usage{
tab_rda(
  tabular,
  tab_filename = NULL,
  rda_filename = NULL,
  hst_filename = NULL,
  explanatory_vars = NULL,
  secret_var = NULL,
  cost_var = NULL,
  hrc = NULL,
  totcode = getOption("rtauargus.totcode"),
  value = NULL,
  freq = NULL,
  maxscore = NULL,
  ip = NULL,
  maxscore_2 = NULL,
  maxscore_3 = NULL,
  decimals = getOption("rtauargus.decimals"),
  hierleadstring = getOption("rtauargus.hierleadstring"),
  codelist = NULL,
  separator = getOption("rtauargus.separator"),
  secret_no_pl = NULL
)
}
\arguments{
\item{tabular}{data.frame which contains the tabulated data and
an additional boolean variable that indicates the primary secret of type boolean \cr
( data.frame contenant les données tabulées et
une variable supplémentaire indiquant le secret primaire de type booléen.)}

\item{tab_filename}{tab file name (with .tab extension) \cr
nom du fichier tab (avec extension .tab)}

\item{rda_filename}{rda file name (with .rda extension) \cr
nom du fichier rda (avec extension)}

\item{hst_filename}{hst file name (with .hst extension) \cr
nom du fichier hst (avec extension)}

\item{explanatory_vars}{Vector of explanatory variables \cr
Variables catégorielles, sous forme  de vecteurs \cr
Example : \code{c("A21", "TREFF", "REG")} for a table crossing
\code{A21} x \code{TREFF} x \code{REG}
(Variable indiquant le secret primaire de type booléen:
prend la valeur "TRUE" quand les cellules du tableau doivent être masquées
par le secret primaire, "FALSE" sinon. Permet de créer un fichier d'apriori)}

\item{secret_var}{Nae of the boolean variable which specifies the secret, primary or not :
equal to "TRUE" if a cell is concerned by the secret,"FALSE" otherwise.
will  be exported in the apriori file. \cr
(Variable indiquant le secret  de type booléen:
prend la valeur "TRUE" quand les cellules du tableau doivent être masquées
"FALSE" sinon. Permet de créer un fichier d'apriori)}

\item{cost_var}{Numeric variable allow to change the cost suppression of a cell
for secondary suppression, it's the value of the cell by default, can be
specified for each cell, fill with NA if the cost doesn't need to be changed
for all cells \cr
(Variable numeric qui permet de changer la coût de suppression d'une cellule,
pris en compte dans les algorithmes de secret secondaire.Par défaut le coût
correspond à la valeur de la cellule.  peut être spécifié pour chacune des cellules,
peut contenir des NA pour les coûts que l'on ne souhaite pas modifier.)
(nombre minimal de décimales à afficher (voir section 'Number of decimals').)}

\item{hrc}{Informations of hierarchical variables (see section
'Hierarchical variables'). \cr
(Informations sur les variables hiérarchiques (voir section
'Hierarchical variables').)
(Caractère qui, répété n fois, indique que la valeur est
à n niveaux de profondeur dans la hiérarchie.)}

\item{totcode}{Code(s) which represent the total of a categorical variable
(see section 'Specific parameters' for this parameter's syntax).
If unspecified for a variable(neither by default nor explicitly)
it will be set to  \code{rtauargus.totcode}. \cr
(Code(s) pour le total d'une variable catégorielle (voir
section 'Specific parameters' pour la syntaxe de ce paramètre). Les
variables non spécifiées (ni par défaut, ni explicitement) se verront
attribuer la valeur de \code{rtauargus.totcode}.)}

\item{value}{Name of the column containing the value of the cells. \cr
(Nom de la colonne contenant la valeur des cellules)}

\item{freq}{Name of the column containing the cell frequency. \cr
(Nom de la colonne contenant les effectifs pour une cellule)}

\item{maxscore}{Name of the column containing, the value of the largest
contributor of a cell. \cr
(Nom de la colonne contenant la valeur du plus gros contributeur
d'une cellule)}

\item{ip}{Value of the safety margin in \% (must be an integer).
(Valeur pour les intervalles de protection en \%, doit être entier )}

\item{maxscore_2}{Name of the column containing, the value of the second largest
contributor to a cell. \cr
(Nom de la colonne contenant la valeur du deuxième plus gros contributeur
d'une cellule)}

\item{maxscore_3}{Name of the column containing, the value of the third largest
contributor to a cell. \cr
(Nom de la colonne contenant la valeur du troisième plus gros contributeur
d'une cellule)}

\item{decimals}{Minimum number of decimals to display
(see section 'Number of decimals') \cr}

\item{hierleadstring}{The character that is used to indicate the depth of a
code in the hierarchy. \cr}

\item{codelist}{file(s) containing labels of a categorical variables
(see section 'Specific parameters' for the syntax of this parameter). \cr
(Fichier(s) contenant les libellés des variables catégorielles
(voir section 'Specific parameters' pour la syntaxe de ce paramètre).)}

\item{separator}{Character used as separator in the .tab file. \cr
(Caractère utilisé en tant que separateur dans le fichier .tab)}

\item{secret_no_pl}{name of a boolean variable which indicates the cells
on which the protection levels won't be applied. If \code{secret_no_pl = NULL}
(default), the protection levels are applied on each cell which gets a \code{TRUE}
status for the \code{secret_var}.\cr}
}
\value{
Return the rda file name as a list (invisible).\cr
(Renvoie le nom du fichier rda sous forme de liste (de
manière invisible).)
}
\description{
Creates an apriori file for the primary secret,
a tabular file (tab) and a metadata file (rda)
from tabulated data and additional information.\cr
}
\details{
Crée un fichier d'apriori pour le secret primaire,
un fichier tabular (tab) et un fichier de métadonnées
(rda) à partir de données tabulées et d'informations additionnelles.
}
\section{Apriori file }{


The apriori file (.hst) summarizes for each value of the table
if they are concerned by the primary secret or not.
With this file tau-argus will not need to set the primary secret itself.
The parameter \code{secret_var} indicates the name of the primary secret variable.
If there is the additional boolean variable which indicates the primary secret
in the table (of tabulated data), the function tab_rda will create
an apriori file in a format conforming to tauargus. \cr

Le fichier d'apriori (.hst) récapitule pour chaque valeurs
du tableau si elles sont concernées par le secret primaire ou non.
Avec ce fichier tau-argus n'aura plus besoin de poser le secret primaire lui même,
il se basera sur le fichier d'apriori pour le faire.
Le paramètre \code{secret_var} indique le nom de la variable du secret primaire.
Si l'on rajoute cette variable supplémentaire indiquant
le secret primaire (de type booléen) au tableau de données tabulées, la fonction
tab_rda permet de créer un fichier d'apriori au format conforme pour tauargus.
}

\section{Specific parameters}{


The parameters \code{totcode}, and \code{codelist}
must be given in the form of a vector indicating the value to take for each variable.
The names of the elements of the vector give the variable concerned and
the elements of the vector give the value of the parameter for Tau-Argus.
An unnamed element will set the default value for each variable. \cr

(Les paramètres \code{totcode},  et \code{codelist}
sont à renseigner sous la forme d'un vecteur indiquant la valeur à prendre
pour chaque variable.

Les noms des éléments du vecteur donnent la variable concernée, les éléments
du vecteur donnent la valeur du paramètre pour Tau-Argus. Un élément non
nommé constituera la valeur par défaut, qui sera attribuée à toutes les
variables pouvant prendre ce paramètre.)

For example :
\itemize{
\item{\code{totcode = "global"} : writes \verb{<TOTCODE> "global"} for each
explanatory vars}
\item{\code{totcode = c("global", size="total", income="total")} :
\verb{<TOTCODE> "global"} for each variable except for \code{size} and
\code{income}} assigned with \verb{<TOTCODE> "total"}
by default : {\if{html}{\out{<TOTCODE>}} "Total"}
\item{\code{totcode = "global"} : écrit \verb{<TOTCODE> "global"} pour
toutes les variables catégorielles}
\item{\code{totcode = c("global", size="total", income="total")} :
\verb{<TOTCODE> "global"} pour toutes les variables catégorielles
sauf  \code{size} and \code{income}} qui se verront affecter
le total : \verb{<TOTCODE> "total"}
Par defaut : {\if{html}{\out{<TOTCODE>}} "Total"}
}
}

\section{Hierarchical variables}{


Parameter \code{hrc} has the same syntax as \code{totcode} and
\code{codelist} (named vector containing as many elements as variables to describe).
Hierarchy is defined in an separate hrc file (\strong{hiercodelist}).
which can be written with the function \code{link{write_hrc2}}.
The function expects the location of this file (and a possible \code{hierleadstring}
if it differs from the default option of the package : @.
The path to the existing file is explicitly given.
The elements of the vector in parameter must be named (with the name of the variable),
even if there is only one element.

emph{Example :}\code{c(category="category.hrc")} \cr

(Le paramètre \code{hrc} obéit aux mêmes règles de syntaxe que \code{totcode}
et \code{codelist} (vecteur nommé contenant autant d'éléments
que de variables à décrire).

La hiérarchie est définie dans un fichier hrc à part (\strong{hiercodelist})
qui peut être écrit à l'aide de la fonction \code{\link[=write_hrc2]{write_hrc2()}}.

La fonction attend l'emplacement de ce fichier (et un éventuel \code{hierleadstring}
s'il diffère de l'option par défaut du package).
Le chemin vers le fichier existant est explicitement donné.
Les éléments du vecteur en paramètre doivent nommés (avec le nom de la variable),
même s'il n'y a qu'un seul élément.

\emph{Exemple :}\code{c(category="category.hrc")})
}

\section{Number of decimals}{

Parameter \code{decimals} indicates the minimum number of decimal places to
include in the output file
(whatever the number of decimals actually present in \code{tabular}).
It applies to all real variables (double) but not to integer variables.
To add zeros to an integer variable, convert it with \code{as.double} beforehand.\cr

(Le paramètre \code{decimals} indique le nombre minimal de décimales à faire
figurer dans le fichier en sortie (quel que soit le nombre de décimales
effectivement présent dans \code{tabular}). Il s'applique à toutes les
variables réelles (double) mais pas aux variables entières (integer). Pour
ajouter des zéros à une variable entière, la convertir avec \code{as.double}
au préalable.)
}

\examples{
\dontrun{
# donnees fictives

tab <-data.frame(
 category       = c( "A" ,   "B",   "C",   "D",   "E",   "F"),
 size           = c("tr1", "tr3", "tr2", "tr1", "tr1", "tr2"),
 area           = c( "07",  "01",  "04",  "06",  "02",  "06"),
 income         = c(  100,     4,     7,    14,    42,    85),
 freq           = c(    2,     6,     8,    45,   100,     1),
 max            = c(   54,     2,     1,    13,    19,    85),
 primary_secret = c( TRUE, FALSE, FALSE,  TRUE, FALSE,  TRUE),
 cost           = c(   NA,    NA,    NA,     1,     5,    NA)
)

# rda creation

files_names <-
 tab_rda(
  tabular          = tab,
  tab_filename     = "tauargus_files/file.tab",
  rda_filename     = "tauargus_files/file.rda",
  hst_filename     = "tauargus_files/file.hst",
  hrc              = c(category = "category.hrc"),
  explanatory_vars = c("category" , "size", "area"),
  secret_var       = "primary_secret",
  cost_var         = "cost"
  totcode          = c(
    category = "global",
    size     = "total",
    area     = "global"
  ),
  value            = "income",
  freq             = "freq"
)

# Viewing product files
file.show(
 res$rda_filename,
  header = unlist(res),
  pager = "internal"
)
}
}
