% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_rtauargus.R
\name{tab_rtauargus}
\alias{tab_rtauargus}
\title{Title}
\usage{
tab_rtauargus(
  tabular,
  files_name = NULL,
  dir_name = NULL,
  explanatory_vars,
  totcode,
  hrc = NULL,
  secret_var = NULL,
  cost_var = NULL,
  value = "value",
  freq = "freq",
  maxscore = NULL,
  safety_rules = "MAN(10)",
  suppress = "MOD(1,5,1,0,0)",
  show_batch_console = FALSE,
  output_type = 4,
  output_options = "",
  unif_hrc = TRUE,
  unif_expl = TRUE,
  ...
)
}
\arguments{
\item{tabular}{[\strong{obligatoire/mandatory}]
data.frame which contains the tabulated data and
an additional boolean variable that indicates the primary secret of type boolean \cr
data.frame contenant les données tabulées et
une variable supplémentaire indiquant le secret primaire de type booléen}

\item{files_name}{string used to name all the files needed to process.
All files will have the same name, only their extension will be different.}

\item{dir_name}{string indicated the path of the directory in which to save
all the files (.rda, .hst, .txt, .arb, .csv) generated by the function.}

\item{explanatory_vars}{[\strong{obligatoire/mandatory}] Vector of explanatory variables \cr
Variables catégorielles, sous forme  de vecteurs \cr
Example : \code{c("A21", "TREFF", "REG")}
table crossing \code{A21} x \code{TREFF} x \code{REG}}

\item{totcode}{Code(s) which represent the total of a categorical variable
(see section 'Specific parameters' for this parameter's syntax).
If unspecified for a variable(neither by default nor explicitly)
it will be set to  \code{rtauargus.totcode}. \cr
(Code(s) pour le total d'une variable catégorielle (voir
section 'Specific parameters' pour la syntaxe de ce paramètre). Les
variables non spécifiées (ni par défaut, ni explicitement) se verront
attribuer la valeur de \code{rtauargus.totcode}.)#'}

\item{hrc}{Informations of hierarchical variables (see section
'Hierarchical variables'). \cr
(Informations sur les variables hiérarchiques (voir section
'Hierarchical variables').)#'}

\item{secret_var}{Boolean variable which give the primary secret : equal to
"TRUE" if a cell is concerned by the primary secret,"FALSE" otherwise.
will  be exported in the apriori file \cr
(Variable indiquant le secret primaire de type booléen:
prend la valeur "TRUE" quand les cellules du tableau doivent être masquées
par le secret primaire, "FALSE" sinon. Permet de créer un fichier d'apriori)}

\item{cost_var}{Numeric variable allow to change the cost suppression of a cell
for secondary suppression, it's the value of the cell by default, can be
specified for each cell, fill with NA if the cost doesn't need to be changed
for all cells \cr
Variable numeric qui permet de changer la coût de suppression d'une cellule,
pris en compte dans les algorithmes de secret secondaire.Par défaut le coût
correspond à la valeur de la cellule.  peut être spécifié pour chacune des cellules,
peut contenir des NA pour les coûts que l'on ne souhaite pas modifier}

\item{value}{Name of the column containing the value of the cells. \cr
(Nom de la colonne contenant la valeur des cellules)}

\item{freq}{Name of the column containing the cell frequency. \cr
(Nom de la colonne contenant les effectifs pour une cellule)#'}

\item{maxscore}{Name of the column containing, the value of the largest
contributor of a cell. \cr
(Nom de la colonne contenant la valeur du plus gros contributeur
d'une cellule)}

\item{safety_rules}{[\strong{obligatoire/mandatory}]
Rules for primary suppression with Argus syntax, if the primary suppression
has been dealt with an apriori file specify manual safety range :"MAN(10)"
for example.\cr
Règle(s) de secret primaire.
Chaîne de caractères en syntaxe batch Tau-Argus. Si le secret primaire
a été traité dans un fichier d'apriori : utiliser "MAN(10)"}

\item{suppress}{[\strong{obligatoire/mandatory}]
Algortihm for secondary suppression (Tau-Argus batch syntax), and the
parameters for it.
Algorithme de gestion du secret secondaire (syntaxe batch de Tau-Argus).
ainsi que les potentiels paramètres associés}

\item{show_batch_console}{to display the batch progress in the
console. \cr
(pour afficher le déroulement du batch dans la
  console.)}

\item{output_type}{Type of the output file (Argus codification)
By default \code{"2"} (csv for pivot-table).
For SBS files use \code{"4"}
Format des fichiers en sortie (codification Tau-Argus).
Valeur par défaut du package : \code{"2"} (csv for pivot-table).
Pour le format SBS utiliser \code{"4"}}

\item{output_options}{Additionnal parameter for the output,
by default : code{"AS+"} (print Status). To specify no options : \code{""}.
  options supplémentaires des fichiers en sortie. Valeur
  par défaut du package : \code{"AS+"} (affichage du statut). Pour ne
  spécifier aucune option, \code{""}.}

\item{unif_hrc}{boolean, if hrc files have to be standardized (especially for internal use)}

\item{unif_expl}{boolean, if explanatory variables have to be standardized (especially for internal use)}

\item{...}{any parameter of the tab_rda, tab_arb or run_arb functions, relevant
for the treatment of tabular.}
}
\value{
If output_type equals to 4, then the original tabular is returned with a new
column called Status, indicating the status of the cell coming from Tau-Argus :
"A" for a primary secret due to frequency rule, "B" for a primary secret due
to dominance rule, "D" for secondary secret and "V" for no secret cell.

If output_type doesn't equal to 4, then the raw result from tau-argus is returned.
}
\description{
Title
}
\details{

}
\examples{
#DON'T RUN
library(dplyr)
tabular_leg <- table_legumes_sp \%>\% mutate(n_obs = round(n_obs)) \%>\%
select(ACTIVITY, treff, type_leg, tot, n_obs, max, is_secret_prim)
res <- tab_rtauargus(
tabular = tabular_leg,
files_name = "legumes_rouges",
dir_name = "tauargus_files",
explanatory_vars = c("ACTIVITY", "treff", "type_leg"),
hrc = c(ACTIVITY = "legumes.hrc"),
totcode = c(ACTIVITY = "Ensemble", treff = "Ensemble", type_leg="rouges"),
secret_var = "is_secret_prim",
value = "tot",
freq = "n_obs",
output_type = "4",
output_options = "",
is_tabular = TRUE,
show_batch_console = FALSE,
import = FALSE,
separator = ",",
verbose = FALSE,
unif_hrc = TRUE,
unif_expl = TRUE
)
}
